support/download: keep files downloaded without hash
In the situation where the hash is missing from the hash file, the
dl-wrapper downloads the file again and again until the developer
specifies the hash to complete the download step.

To avoid this situation, the freshly-downloaded file is not removed
anymore after a successful download.

The build continues to terminate in error and the file will not be
downloaded again during the next builds.

Next builds will terminate in error until the hash is specified.

Signed-off-by: Gaël PORTAY <gael.portay@savoirfairelinux.com>