qt5base: specify OpenGL/GLESv2 and EGL libdirs
Qt5Gui requires QMAKE_LIBDIR_OPENGL/QMAKE_LIBDIR_OPENGL_ES2 and
QMAKE_LIBDIR_EGL variables to be defined to use variables
${Qt5Gui_OPENGL_LIBRARIES} and ${Qt5Gui_EGL_LIBRARIES}.

When they are undefined, the Qt5Gui_*_LIBRARIES are left empty. As a
consequence, those variables are unusable. QtWebKit 5.212 is a project
that needs them to be built.

The macro _qt5gui_find_extra_libs() [1] uses cmake find_library() [2] to
locate OpenGL/EGL libraries. While cross-compiling, the QMAKE_LIBDIR_*
variables are used to set the path where to search for those libraries.

If the libraries are found, Qt5Gui_*_LIBRARIES are properly set with an
additional prefix (Qt5::Gui_).

If the path is empty (i.e. when QMAKE_LIBDIR_* are not set), libraries
are not found and Qt5Gui_*_LIBRARIES are not set.

Here are some examples (output from qtwebkit 5.212):

	* raspberrypi with rpi-userland:
	-- Qt OpenGL implementation: GLESv2
	-- Qt OpenGL libraries:
	-- Qt EGL libraries: Qt5::Gui_EGL;Qt5::Gui_GLESv2;Qt5::Gui_vchostif;Qt5::Gui_bcm_host;Qt5::Gui_vcos;Qt5::Gui_vchiq_arm

	* PC with Mesa + OpenGL ES/EGL:
	-- Qt OpenGL implementation: GLESv2
	-- Qt OpenGL libraries: Qt5::Gui_GLESv2
	-- Qt EGL libraries: Qt5::Gui_EGL

	* cubieboard2 with sunxi-mali:
	-- Qt OpenGL implementation: GLESv2
	-- Qt OpenGL libraries: Qt5::Gui_GLESv1_CM
	-- Qt EGL libraries: Qt5::Gui_EGL;Qt5::Gui_GLESv2;Qt5::Gui_UMP

[1]: https://github.com/qt/qtbase/blob/v5.11.1/src/gui/Qt5GuiConfigExtras.cmake.in#L76-L157
[2]: https://github.com/qt/qtbase/blob/v5.11.1/src/gui/Qt5GuiConfigExtras.cmake.in#L96-L103

Signed-off-by: Gaël PORTAY <gael.portay@savoirfairelinux.com>