qt5base: add comment for qt5webkit examples
Unlike every other Qt5 modules, the examples of QtWebKit are located in
the dedicated package qt5webkit-examples. This is confusing.

This commit adds a comment to warn the user that QtWebKit examples are
available under a dedicated package. The comment shows up in the case
where the Qt examples is set and qt5webkit-examples is left unset.

Signed-off-by: Gaël PORTAY <gael.portay@savoirfairelinux.com>