configs/raspberrypi3_qt5we: switch to mesa3d
The package rpi-userland has been marked as ancient and deprecated[1].

It contains code using proprietary APIs to interface to the VideoCore
firmware. It has moved to standard linux APIs.

It is recommended to use DRM/KMS and Mesa instead.

This switches to the mesa3d OpenGLES backend instead of the deprecated
one provided by rpi-userland. It implies the load for the vc4-kms-v3d
device-tree overlay.

[1]: https://github.com/raspberrypi/userland/commit/96a7334ae9d5fc9db7ac92e59852377df63f1848

Signed-off-by: Gaël PORTAY <gael.portay+rtone@gmail.com>